Move Insert and Delete to mathics.list.eol and..
create mathics.eval.lists
