+++ mathics/algorithm/__init__.py
@@ -1,2 +1,5 @@
 #!/usr/bin/env python3
 # -*- coding: utf-8 -*-
+"""
+Note: functions here will eventually get moved to mathics.eval
+"""
