+++ mathics/eval/__init__.py
@@ -1,8 +1,9 @@
 """
 Mathics Evaluation Functions
 
-Routines here are core operations or functions that implement evaluation. If there
-were an instruction interpreter, these would be the instructions.
+Routines here are core operations or functions that implement
+evaluation. If there were an instruction interpreter, these functions
+that start "eval_" would be the interpreter instructions.
 
 These operatations then should include the most commonly-used Builtin-functions like
 ``N[]`` and routines in support of performing those evaluation operations/instructions.
@@ -11,11 +12,12 @@ Performance of the operations here can be important for overall interpreter perf
 
 It may be even be that some of the functions here should be written in faster
 language like C, Cython, or Rust.
+
 """
 
-# Ideally, this module should depend on modules inside ``mathics.core`` but not in modules stored in ``mathics.builtin`` to avoid circular references.
-
+# This module should not depend  on ``mathics.builtin``. Dependence goes only the other way around
 
 # ``evaluation``, ``_rewrite_apply_eval_step``, ``set`` that in the current implementation
 # requires to introduce local imports.
-# This also would make easier to test and profile classes that store Expression-like objects and methods that produce the evaluation.
+
+# Moving evaluation routines out of builtins allows us to test and profile code here.
